Creacion de utilidades, services, repository y controllers para login y registro de usuarios, admin y empleados
creacion de services, repository y controllers para vehiculo y categoria de vehiculos

clase mapper para mapear dtos y entidades

+++ Backend/alquiler_carros/src/main/java/base_datos/alquiler_carros/Repository/IVehiculoRepository.java
@@ -13,11 +13,11 @@ public interface IVehiculoRepository extends JpaRepository<Vehiculo, Integer> {
     Optional<Vehiculo> findByPlaca(String placa);
     boolean existsByPlaca(String placa);
     List<Vehiculo> findByEstado(String estado);
-    List<Vehiculo> findByCategoria_IdCategoria(Integer idCategoria);
-    List<Vehiculo> findBySucursal_IdSucursal(Integer idSucursal);
+    List<Vehiculo> findByCategoria_IdCategoria(Long idCategoria);
+    List<Vehiculo> findBySucursal_IdSucursal(Long idSucursal);
     List<Vehiculo> findByMarca(String marca);
 
     // Vehículos disponibles en una sucursal por categoría
     List<Vehiculo> findByEstadoAndSucursal_IdSucursalAndCategoria_IdCategoria(
-            String estado, Integer idSucursal, Integer idCategoria);
+            String estado, Long idSucursal, Integer idCategoria);
 }
